New version of CFGVechiles with desired unbanned classes.
Now, bear in mind that I'm still with the guys who think we should unban everything and let server owners worry about what they want it and what not and deal with security.

I tried to leave out those vehicles which (by experience) I find to be extremely overpowered or otherwise disruptive to DayZ gameplay.

You might think that some are too powerful, but they have been countered by others that can deal with them, so there's never a "king of the hill" vehicle that can trump all the others. Specially if we take into consideration our new weapon additions.